GERTiMNC v1.3 - Build 3
-- Change: GERTiMNC now solely resides as an RC service in /etc/rc.d

--Change: Configuration files have moved to /etc

-- Feature Addition: triggering the loadAddress command will have the MNC reload the list of addresses. This way, you can change the modem address tied to a certain GERTi address without restarting the entire network. The GERTiClient for that node must be restarted for it to take effect.
